+++ src/Tests/Mvp24Hours.Infrastructure.Data.EFCore.Test/RepositoryBaseTest.cs
@@ -29,7 +29,7 @@ public class RepositoryBaseTest
 
         var repository = new TestableRepositoryBase(context, null!);
 
-        repository.ExposedOptions.MaxQtyByQueryPage.Should().Be(ContantsHelper.Data.MaxQtyByQueryPage);
+        repository.ExposedOptions.MaxQtyByQueryPage.Should().Be(ConstantsHelper.Data.MaxQtyByQueryPage);
         repository.ExposedOptions.TransactionIsolationLevel.Should().BeNull();
     }